linux: Avoid infinite recursion in the tests
The `tests::itest(long)` would call itself instead of calling the
unsigned long long variant as intended.

+++ src/tests.cc
@@ -7080,9 +7080,9 @@ tests &tests::itest(long value)
 // ----------------------------------------------------------------------------
 {
     if (value < 0)
-        return itest(-value, CHS);
+        return itest((unsigned long long) (-value), CHS);
     else
-        return itest(value);
+        return itest((unsigned long long) value);
 }
